+++ CHANGELOG.md
@@ -15,6 +15,8 @@
 - 매직펜 하이파이 디자인 추가(블러효과)
 - 권한 거부에 따른 UI 추가
 - 하드웨어 버튼을 통한 촬영
+- 프레임 비율 조정 시, 프레임 내 그리드와 디밍 효과 추가
+- 양방향 프레임의 수정 점유권에 따른 하이파이 구현
 
 ### Changed
 
@@ -34,6 +36,8 @@
 
 ### Removed
 
+- FrameDisplayView
+- Frame의 Color 프로퍼티
 
 ### Fixed
 
@@ -41,6 +45,7 @@
 - PhotosPickerView에서 썸네일 이미지를 레퍼런스 이미지로 선택하면 정방형 이미지가 로드되던 문제
 - PhotosPickerView와 PhotoDetailView 체크박스 상태 동기화 이슈 해결
 - 레퍼런스 삭제시 PhotoPicker 초기화가 안된 문제 해결
+-  프레임의 동시 수정 관련된 문제 해결
 - SampleBufferDisplayView에서 매번 화질 관련 지표를 리셋하여 적응형 화질 매커니즘이 동작하지 않았던 문제
 
 ### Security